+++ implementations/rust/ockam/ockam_command/src/policy/delete.rs
@@ -1,7 +1,10 @@
 use crate::policy::policy_path;
-use crate::util::{extract_address_value, node_rpc, Rpc};
-use crate::CommandGlobalOpts;
+
+use crate::util::{node_rpc, parse_node_name, Rpc};
+use crate::{fmt_ok, CommandGlobalOpts};
 use clap::Args;
+use colorful::Colorful;
+
 use ockam::Context;
 use ockam_abac::{Action, Resource};
 use ockam_core::api::Request;
@@ -16,6 +19,10 @@ pub struct DeleteCommand {
 
     #[arg(short, long)]
     action: Action,
+
+    /// Confirm the deletion without prompting
+    #[arg(display_order = 901, long, short)]
+    yes: bool,
 }
 
 impl DeleteCommand {
@@ -36,10 +43,22 @@ async fn run_impl(
     opts: CommandGlobalOpts,
     cmd: DeleteCommand,
 ) -> miette::Result<()> {
-    let node = extract_address_value(&cmd.at)?;
-    let req = Request::delete(policy_path(&cmd.resource, &cmd.action));
-    let mut rpc = Rpc::background(ctx, &opts, &node)?;
-    rpc.request(req).await?;
-    rpc.is_ok()?;
+    let node = parse_node_name(&cmd.at)?;
+    if opts
+        .terminal
+        .confirmed_with_flag_or_prompt(cmd.yes, "Are you sure you want to delete this policy?")?
+    {
+        let req = Request::delete(policy_path(&cmd.resource, &cmd.action));
+        let mut rpc = Rpc::background(ctx, &opts, &node)?;
+        rpc.request(req).await?;
+        rpc.is_ok()?;
+
+        opts.terminal
+            .stdout()
+            .plain(fmt_ok!("Policy with name '{}' has been deleted", &cmd.at))
+            .machine(&cmd.at)
+            .json(serde_json::json!({ "policy": { "at": &cmd.at } }))
+            .write_line()?;
+    }
     Ok(())
 }
